Move some types so they are available outside the package
